Clear neopixels on startup (using SPI)
